Application refactor
