+++ v2/internal/app/app_debug.go
@@ -1,5 +1,4 @@
 //go:build debug
-// +build debug
 
 package app
