+++ v2/internal/app/app_debug_not.go
@@ -1,5 +1,4 @@
 //go:build !debug
-// +build !debug
 
 package app
