+++ v2/internal/app/app_default_unix.go
@@ -1,5 +1,4 @@
-//go:build !dev && !production && !bindings && linux
-// +build !dev,!production,!bindings,linux
+//go:build !dev && !production && !bindings && darwin
 
 package app
 
@@ -9,14 +8,9 @@ import (
 	"github.com/wailsapp/wails/v2/pkg/options"
 )
 
-// App defines a Wails application structure
-type App struct{}
-
 func (a *App) Run() error {
 	return nil
 }
-
-func (a *App) Shutdown() {}
 
 // CreateApp creates the app!
 func CreateApp(_ *options.App) (*App, error) {
