+++ v2/internal/app/app_dev.go
@@ -1,5 +1,4 @@
 //go:build dev
-// +build dev
 
 package app
 
@@ -14,7 +13,6 @@ import (
 	"path/filepath"
 
 	"github.com/wailsapp/wails/v2/internal/binding"
-	"github.com/wailsapp/wails/v2/internal/frontend"
 	"github.com/wailsapp/wails/v2/internal/frontend/desktop"
 	"github.com/wailsapp/wails/v2/internal/frontend/devserver"
 	"github.com/wailsapp/wails/v2/internal/frontend/dispatcher"
@@ -26,27 +24,6 @@ import (
 	pkglogger "github.com/wailsapp/wails/v2/pkg/logger"
 	"github.com/wailsapp/wails/v2/pkg/options"
 )
-
-// App defines a Wails application structure
-type App struct {
-	frontend frontend.Frontend
-	logger   *logger.Logger
-	options  *options.App
-
-	menuManager *menumanager.Manager
-
-	// Indicates if the app is in debug mode
-	debug bool
-
-	// OnStartup/OnShutdown
-	startupCallback  func(ctx context.Context)
-	shutdownCallback func(ctx context.Context)
-	ctx              context.Context
-}
-
-func (a *App) Shutdown() {
-	a.frontend.Quit()
-}
 
 func (a *App) Run() error {
 	err := a.frontend.Run(a.ctx)
