+++ v2/internal/app/app_preflight_unix.go
@@ -1,0 +1,12 @@
+//go:build unix && !bindings
+
+package app
+
+import (
+	"github.com/wailsapp/wails/v2/internal/logger"
+	"github.com/wailsapp/wails/v2/pkg/options"
+)
+
+func PreflightChecks(_ *options.App, _ *logger.Logger) error {
+	return nil
+}
