+++ v2/internal/app/app_preflight_windows.go
@@ -1,5 +1,4 @@
 //go:build windows && !bindings
-// +build windows,!bindings
 
 package app
